requirements.txt
python-dotenv==0.19.2
pymongo==3.11.4
pymongo[srv]
jsonschema==3.2.0
